Moved rainfall tests into test_precipitation
