+++ tests/operators/test_precipitation.py
@@ -14,12 +14,16 @@
 
 """Test precipitation operators."""
 
+from pathlib import Path
+
 import cf_units
 import iris.cube
 import numpy as np
 import pytest
 
 from CSET.operators import precipitation
+
+DATA_DIR = Path(__file__).resolve().parent.parent / "test_data" / "rainfall"
 
 
 def test_maul_properties_wrong_output(maul_mask):
@@ -259,3 +263,62 @@ def test_maul_properties_5D_depth(maul_mask_all, precalc_maul_depth_5d):
         atol=1e-6,
         equal_nan=True,
     )
+
+
+def test_rain_amount_to_rate_with_bounds():
+    """Test conversion works with time bounds."""
+    cube_in = iris.load_cube(DATA_DIR / "rain_amount_bounds.nc")
+    cube = cube_in.copy()
+    out = precipitation.convert_rainfall_amount_to_rate(cube)
+    assert out.units.is_convertible("kg m-2 s-1")
+    expected = np.asarray(cube_in.data, dtype=float) / 1800.0
+    assert np.allclose(out.data, expected, rtol=0.0, atol=1e-12)
+
+
+def test_rain_amount_to_rate_without_bounds_uses_time_points():
+    """Test conversion works without time bounds."""
+    cube_in = iris.load_cube(DATA_DIR / "rain_amount_no_bounds.nc")
+    cube = cube_in.copy()
+    out = precipitation.convert_rainfall_amount_to_rate(cube)
+    assert out.units.is_convertible("kg m-2 s-1")
+    expected = np.asarray(cube_in.data, dtype=float) / 1800.0
+    assert np.allclose(out.data, expected, rtol=0.0, atol=1e-12)
+
+
+def test_rain_rate_is_left_untouched():
+    """Test that rainfall rate is left untouched."""
+    cube_in = iris.load_cube(DATA_DIR / "rain_rate.nc")
+    cube = cube_in.copy()
+    out = precipitation.convert_rainfall_amount_to_rate(cube)
+    assert out.units == cube_in.units
+    assert np.allclose(out.data, cube_in.data, rtol=0.0, atol=0.0)
+
+
+def test_non_rainfall_units_are_skipped():
+    """Test that non-rainfall units are skipped."""
+    cube_in = iris.load_cube(DATA_DIR / "not_rainfall_units.nc")
+    cube = cube_in.copy()
+    out = precipitation.convert_rainfall_amount_to_rate(cube)
+    assert out.units == cube_in.units
+    assert np.allclose(out.data, cube_in.data, rtol=0.0, atol=0.0)
+
+
+def test_raises_if_no_time_coordinate():
+    """Test that error raised if no time coordinate."""
+    cube = iris.load_cube(DATA_DIR / "no_time.nc")
+    with pytest.raises(ValueError, match="No time coordinate"):
+        precipitation.convert_rainfall_amount_to_rate(cube)
+
+
+def test_raises_if_single_time_point_no_bounds():
+    """Test that error raised if only one time available."""
+    cube = iris.load_cube(DATA_DIR / "single_time.nc")
+    with pytest.raises(ValueError, match="single time point"):
+        precipitation.convert_rainfall_amount_to_rate(cube)
+
+
+def test_raises_if_nonpositive_interval_from_bounds():
+    """Test that error raised if time interval is negative."""
+    cube = iris.load_cube(DATA_DIR / "bad_bounds_zero.nc")
+    with pytest.raises(ValueError, match="Non-positive rainfall accumulation interval"):
+        precipitation.convert_rainfall_amount_to_rate(cube)
